# pkg/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go API Template",
        "title": "Go API Template",
        "contact": {
            "name": "Chenchuan.Song",
            "email": "robscc269@gmail.com"
        },
        "license": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/sample/post_file_handler": {
            "post": {
                "description": "method sample post file",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Method"
                ],
                "summary": "PostFileHandler",
                "parameters": [
                    {
                        "type": "file",
                        "description": "the file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sample/post_handler": {
            "post": {
                "description": "method sample post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Method"
                ],
                "summary": "PostHandler",
                "parameters": [
                    {
                        "description": "PostSampleParams the sample_key is required",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/controller.PostSampleParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.PostSampleParams": {
            "type": "object",
            "required": [
                "sample_Key"
            ],
            "properties": {
                "sample_Key": {
                    "description": "sample key is sample and must be valued when post to server",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}